+++ include/common.h
@@ -4,13 +4,14 @@
 #define SUBDATANUM 2000000
 #define DATANUM (SUBDATANUM * MAX_THREADS)   /* total number of data */
 
-typedef enum {
+enum Mode {
+    LOCAL,
     CLIENT,
     SERVER
-} Mode;
+};
 
 void merge(float arr[], const float left[], const float right[], const int leftSize, const int rightSize);
 void mergeSort(const float data[], const int size, float result[]);
 
 void run_original(const float data[], const int len, float& sum_value, float& max_value, float result[]);
-void run_speedup(const float data[], const int len, const Mode mode, float& sum_value, float& max_value, float result[]);
+void run_speedup(const float data[], const int len, float& sum_value, float& max_value, float result[]);